requirements.txt
httpx>=0.27
pydantic>=2.0
pydantic-settings>=2.0
python-telegram-bot>=21.0
apscheduler>=3.10
pyyaml>=6.0
feedparser>=6.0
